pass tests

+++ src/sessions/session.controller.ts
@@ -37,15 +37,21 @@ export class SessionController {
       return
     }
     if (session.attended) {
-      res.status(405).send('Cannot update an attended session.')
+      res.status(409).send('Conflict: Cannot update an attended session.')
+      return
     }
     session = await this.repository.update({ id }, data)
     return session
   }
-  @Delete('/')
+  @Delete('/:id')
   @Status(200)
   async delete(req: Request, res: Response) {
     const { id } = ResourceIdentifier.parse(req.params)
+    const session = await this.repository.readOne({ id })
+    if(!session){
+      res.sendStatus(404)
+      return 
+    }
     await this.repository.delete({ id })
     res.sendStatus(200)
   }
